requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-sqlite
pyspark==3.5.0
kafka-python==2.0.2
cassandra-driver==3.29.3
requests==2.31.0
pytz==2023.3